Revision 4
Condensed classes and updated references accordingly
