+++ src/ijordan/matrixonatorTest/MatrixOperationsTest.java
@@ -2,18 +2,51 @@ package ijordan.matrixonatorTest;
 
 import static org.junit.Assert.*;
 import ijordan.matrixonator.model.Matrix;
-import ijordan.matrixonator.model.MatrixSolver;
-import ijordan.matrixonator.model.SimpleMatrixArithmetic;
+import ijordan.matrixonator.model.MatrixOperations;
 
 import org.junit.Test;
 
+public class MatrixOperationsTest {
 
+	/* --------------------------
+	 *  Matrix Solving Tests
+	 */
+	
+	@Test
+	public void testMatrixReduceNormal() {
+		double[][] data = { { 1, 0, 3, 3 }, { 0, 1, 0, 4 }, { 0, 0, 0, 1 } };
+		double[][] dataResult = { { 1, 0, 3, 0 }, { 0, 1, 0, 0 }, { 0, 0, 0, 1 } };
+		final Matrix testMatrix = new Matrix("Test1", data);
 
-public class SimpleMatrixArithmeticTest {
+		assertEquals("Matrix Reduction gives incorrect result", dataResult,
+									MatrixOperations.reducedEchelonForm(testMatrix).getData());
+	}
 
-	//Required solver and arithmetic (Since Revision 2)
-	private SimpleMatrixArithmetic SMA = new SimpleMatrixArithmetic();
-	private MatrixSolver MS = new MatrixSolver();
+	@Test
+	public void testMatrixReduceNegative() {
+		double[][] data = { { 1, 2, 1 }, { -2, -3, 1 }, { 3, 5, 0 } };
+		double[][] dataResult = { { 1, 0, -5 }, { 0, 1, 3 }, { 0, 0, 0 } };
+		final Matrix testMatrix = new Matrix("Test2", data);
+
+		assertEquals("Matrix Reduction gives incorrect result", dataResult, 
+													MatrixOperations.reducedEchelonForm(testMatrix).getData());
+	
+	}
+
+	@Test
+	public void testMatrixReduceLargeValue() {
+		double[][] data = { { 1586, 7546, 85746, 57564 }, { 756, 374, 385, 0 },
+				{ 8765, 123, 765, 234 }, { 5736, 4736, 27364, 5 } };
+		double[][] dataResult = { { 1, 0, 0, 0 }, { 0, 1, 0, 0 }, { 0, 0, 1, 0 }, { 0, 0, 0, 1 } };
+		final Matrix testMatrix = new Matrix("Test2", data);
+
+		assertEquals("Matrix Reduction gives incorrect result", dataResult, MatrixOperations.reducedEchelonForm(testMatrix).getData());
+	}
+
+	
+	/* ------------
+	 * Matrix Arithmetic Tests
+	 */
 	
 	@Test
 	public void testMatrixMultiplication1() {
@@ -27,10 +60,10 @@ public class SimpleMatrixArithmeticTest {
 		final Matrix testMatrixResult = new Matrix("TestResult", dataResult);
 
 		assertTrue("Matrices should be compatible",
-				SMA.checkMultCompatibility(testMatrix1, testMatrix2));
+				MatrixOperations.checkMultCompatibility(testMatrix1, testMatrix2));
 
 		assertEquals("Matrix Multiplication gives incorrect result", testMatrixResult.getData(),
-				SMA.multiplyMatrices(testMatrix1, testMatrix2).getData());
+				MatrixOperations.multiplyMatrices(testMatrix1, testMatrix2).getData());
 
 	}
 
@@ -49,10 +82,10 @@ public class SimpleMatrixArithmeticTest {
 		final Matrix testMatrixResult = new Matrix("TestResult", dataResult);
 
 		assertTrue("Matrices should be compatible",
-				SMA.checkMultCompatibility(testMatrix1, testMatrix2));
+				MatrixOperations.checkMultCompatibility(testMatrix1, testMatrix2));
 
 		assertEquals("Matrix Multiplication gives incorrect result", testMatrixResult.getData(),
-				SMA.multiplyMatrices(testMatrix1, testMatrix2).getData());
+				MatrixOperations.multiplyMatrices(testMatrix1, testMatrix2).getData());
 
 	}
 
@@ -65,10 +98,10 @@ public class SimpleMatrixArithmeticTest {
 		final Matrix testMatrix2 = new Matrix("Test2", dataSecond);
 
 		assertFalse("Matrices should not be compatible",
-				SMA.checkMultCompatibility(testMatrix1, testMatrix2));
+				MatrixOperations.checkMultCompatibility(testMatrix1, testMatrix2));
 
 		assertFalse("Matrices should not be compatible",
-				SMA.checkMultCompatibility(testMatrix2, testMatrix1));
+				MatrixOperations.checkMultCompatibility(testMatrix2, testMatrix1));
 
 	}
 
@@ -86,9 +119,9 @@ public class SimpleMatrixArithmeticTest {
 		final Matrix testMatrix2 = new Matrix("Test2", dataSecond);
 		final Matrix testMatrixResult = new Matrix("TestResult", dataResult);
 
-		assertEquals("Matrix Addition gives incorrect result", testMatrixResult.getData(), SMA
+		assertEquals("Matrix Addition gives incorrect result", testMatrixResult.getData(), MatrixOperations
 				.addMatrices(testMatrix1, testMatrix2).getData());
 
 	}
-
+	
 }
